+++ SharePointFramework/ProjectWebParts/src/loc/en-us.js
@@ -2,6 +2,7 @@ define([], function () {
   return {
     EditTimelineContentText: "Edit timeline item",
     NewTimelineContentText: "New timeline item",
+    TimelineEndDateMilestoneDescription: "End date is required for element types displayed as a point on the timeline ({0})",
     NewStatusPanelTitle: "New status report",
     EditStatusPanelText: "Edit status report",
     ErrorBoundaryFallbackDescription: "Error details:",