+++ SharePointFramework/ProjectWebParts/src/loc/mystrings.d.ts
@@ -2,6 +2,7 @@ declare interface IProjectWebPartsStrings {
   OpportunityMatrixConfigurationFolder: string
   EditTimelineContentText: string
   NewTimelineContentText: string
+  TimelineEndDateMilestoneDescription: string
   NewStatusPanelTitle: string
   EditStatusPanelText: string
   Aria: {